Traktor: Scale factor added to mesh asset. Fixed unaligned reference compile error on x86 builds. Minor type fix in ActionFunction dispatchers.
[git-p4: depot-paths = "//depot/traktor/": change = 19352]

+++ code/Flash/Sound/FlashSoundBuffer.cpp
@@ -58,7 +58,7 @@ bool FlashSoundBuffer::getBlock(sound::ISoundBufferCursor* cursor, const sound::
 
 	// Have we reached the end?
 	int32_t position = fsbc->m_position;
-	if (position >= m_sound->getSamplesCount())
+	if (position >= int32_t(m_sound->getSamplesCount()))
 		return false;
 
 	// Calculate number of samples to convert for this block.
